backend/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-jwt-extended>=4.6
supabase>=2.0
python-dotenv>=1.0
marshmallow>=3.20
